requirements.txt
grpcio==1.21.1
grpcio-tools==1.21.1
protobuf==3.8.0
six==1.12.0
